Support managed database connections with TLS
When using managed databases solutions by cloud providers the database
is running in a remote server. The connection is secured with TLS, thus
we need to configure explicitly the certificate authority in the
database connector.

The changes here add that possibility optionally. It is an optional
configuration that can be used on demand. OpenSSL is already an
dependency from other sub-crates, so I imagine this won't add much to
the compilation time.